# optimized line fit for 2d case by upto 2665.96% by using closed form solution