Implemented the 2d case for the fit line using the standard OLS algorithm, for structured data it works perfectly as the original svd implementation, but in cases where the data is a random point cloud it may differ but that makes sense as any possible answer would be technically correct in those cases.

=== Benchmark: Weighted Case ===
Weighted OLS time: 1.06 ms
SVD time:          21.48 ms
Speedup:           **2019.57%**
Cosine similarity mean: 1.000000
Max angular error (deg): 0.034264
Mean angular error (deg): 0.000459
Number of batches with >5° error: 0

=== Benchmark: Unweighted Case ===
Unweighted OLS time: 0.62 ms
SVD time:            16.53 ms
Speedup:             **2665.96%**
Cosine similarity mean: 1.000000
Max angular error (deg): 0.019782
Mean angular error (deg): 0.000320
Number of batches with >5° error: 0

https://colab.research.google.com/drive/1pDMN_xZzb6GVmA7u0mPR87eo-FY2AFL6?usp=sharing

#### Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] Did you update CHANGELOG in case of a major change?